Install 32bit gtk too

+++ .github/install_gtk.cmd
@@ -1,8 +1,16 @@
+if ( Test-Path -Path 'D:\a\_temp\Library\RGtk2\gtk' -PathType Container ) { exit 0 }
+
 echo "Installing RGtk2 ..."
 Rscript -e "install.packages('RGtk2', repos = 'https://cloud.r-project.org')"
 
 echo "Installing gtk ..."
 curl -L -o gtk.zip http://ftp.gnome.org/pub/gnome/binaries/win64/gtk+/2.22/gtk+-bundle_2.22.1-20101229_win64.zip
+md gtk
+7z x gtk.zip -ogtk > nul
+REM del gtk.zip
+mv gtk D:\a\_temp\Library\RGtk2\
+
+curl -L -o gtk.zip http://ftp.gnome.org/pub/gnome/binaries/win32/gtk+/2.22/gtk+-bundle_2.22.1-20101227_win32.zip
 md gtk
 7z x gtk.zip -ogtk > nul
 REM del gtk.zip
